# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/logout": {
            "get": {
                "description": "Logout\nDelete the OAuth Bearer Token provided in the Authorization header (e.g. to log out).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (any role)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token that was deleted",
                        "schema": {
                            "$ref": "#/definitions/token.Token"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
...
        "/v1/token_user_ids": {
            "get": {
                "description": "List User IDs for which tokens exist\nList User IDs for which tokens exist. This is useful for paging through tokens by user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (Administrator)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Reverse Order (default: false)",
                        "name": "reverse",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit (default: 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Offset (default: forward/reverse alphanumeric)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated User IDs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid query parameter)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Unauthorized (not an Administrator)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/tokens": {
            "get": {
                "description": "List Tokens\nList OAuth Bearer Tokens by User ID (defaults to the Context User).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (any role)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User ID or Email (defaults to the Context User)",
                        "name": "user",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tokens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/token.Token"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid User ID or Email)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Forbidden (only administrators may read any user's tokens)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new Token\nCreate a new OAuth Bearer Token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "description": "Token Request",
                        "name": "TokenRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/token.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Token Response",
                        "schema": {
                            "$ref": "#/definitions/token.Response"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "URL of the newly created Token"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid JSON body)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthorized (invalid username or password)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            }
        },
        "/v1/tokens/{id}": {
            "get": {
                "description": "Read Token\nRead OAuth Bearer Token by ID (e.g. to verify a token has not expired).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (any role)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token",
                        "schema": {
                            "$ref": "#/definitions/token.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Forbidden (only administrators may read any user's tokens)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found (Token not found)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Token\nDelete OAuth Bearer Token by ID (e.g. to log out).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth Bearer Token (any role)",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token that was deleted",
                        "schema": {
                            "$ref": "#/definitions/token.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "401": {
                        "description": "Unauthenticated (missing or invalid Authorization header)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "403": {
                        "description": "Forbidden (only administrators may delete any user's tokens)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "404": {
                        "description": "Not Found (Token not found)",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.APIEvent"
                        }
                    }
                }
            },
            "head": {
                "description": "Token Exists\nCheck if the specified OAuth Bearer Token exists.",
                "tags": [
                    "Token"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Token Exists"
                    },
                    "400": {
                        "description": "Bad Request (invalid path parameter ID)"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.APIEvent": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "eventID": {
                    "type": "string"
                },
                "logLevel": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
        "token.Request": {
            "type": "object",
            "properties": {
                "grantType": {
                    "description": "usually \"password\"",
                    "type": "string"
                },
                "password": {
                    "description": "plaintext password",
                    "type": "string"
                },
                "username": {
                    "description": "email or User ID",
                    "type": "string"
                }
            }
        },
        "token.Response": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "description": "Token ID (a tuid.TUID)",
                    "type": "string"
                },
                "expiresAt": {
                    "description": "when the token expires (DynamoDB TTL)",
                    "type": "string"
                },
                "tokenType": {
                    "description": "usually \"Bearer\"",
                    "type": "string"
                }
            }
        },
        "token.Token": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    }
}